have reviews related through books model, need to switch
